feat: add expiration select to update api-key
