+++ src/routes/console/project-[project]/overview/keys/[key]/+page@project-[project].svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
     import { invalidate } from '$app/navigation';
+    import { trackEvent } from '$lib/actions/analytics';
     import { CardGrid, Heading, Secret } from '$lib/components';
     import { Dependencies } from '$lib/constants';
-    import { Button, Form, FormList, InputText, InputDateTime } from '$lib/elements/forms';
+    import { Button, Form, FormList, InputText } from '$lib/elements/forms';
     import { symmetricDifference } from '$lib/helpers/array';
     import { toLocaleDateTime } from '$lib/helpers/date';
     import { Container } from '$lib/layout';
@@ -10,21 +11,19 @@
     import { sdkForConsole } from '$lib/stores/sdk';
     import { onMount } from 'svelte';
     import { project } from '../../../store';
-    import { key } from './store';
     import Scopes from '../scopes.svelte';
     import Delete from './delete.svelte';
-    import { trackEvent } from '$lib/actions/analytics';
+    import { key } from './store';
+    import UpdateExpirationDate from './updateExpirationDate.svelte';
 
     let showDelete = false;
     let name: string = null;
     let secret: string = null;
-    let expire: string = null;
     let scopes: string[] = null;
 
     onMount(() => {
         name ??= $key.name;
         secret ??= $key.secret;
-        expire ??= $key.expire;
         scopes ??= $key.scopes;
     });
 
@@ -42,29 +41,6 @@
             addNotification({
                 type: 'success',
                 message: 'API Key name has been updated'
-            });
-        } catch (error) {
-            addNotification({
-                type: 'error',
-                message: error.message
-            });
-        }
-    }
-
-    async function updateExpire() {
-        try {
-            await sdkForConsole.projects.updateKey(
-                $project.$id,
-                $key.$id,
-                $key.name,
-                $key.scopes,
-                expire
-            );
-            invalidate(Dependencies.KEY);
-            trackEvent('submit_key_update_expire');
-            addNotification({
-                type: 'success',
-                message: 'API Key expiration has been updated'
             });
         } catch (error) {
             addNotification({
@@ -165,21 +141,9 @@
             </svelte:fragment>
         </CardGrid>
     </Form>
-    <Form on:submit={updateExpire}>
-        <CardGrid>
-            <Heading tag="h6" size="7">Update Expiration Date</Heading>
-            <p class="text">Choose any name that will help you distinguish between API keys.</p>
-            <svelte:fragment slot="aside">
-                <FormList>
-                    <InputDateTime id="expire" label="Expiration Date" bind:value={expire} />
-                </FormList>
-            </svelte:fragment>
 
-            <svelte:fragment slot="actions">
-                <Button disabled={expire === $key.expire} submit>Update</Button>
-            </svelte:fragment>
-        </CardGrid>
-    </Form>
+    <UpdateExpirationDate />
+
     <CardGrid danger>
         <div>
             <Heading tag="h6" size="7">Delete API Key</Heading>
