+++ src/routes/console/project-[project]/overview/keys/wizard/step1.svelte
@@ -1,31 +1,8 @@
 <script lang="ts">
-    import { FormList, InputText, InputDateTime, InputSelect } from '$lib/elements/forms';
+    import { FormList, InputText } from '$lib/elements/forms';
     import { WizardStep } from '$lib/layout';
+    import ExpirationInput from '../expirationInput.svelte';
     import { key } from './store';
-
-    function incrementToday(value: number, type: 'day' | 'month' | 'year'): string {
-        const date = new Date();
-        switch (type) {
-            case 'day':
-                date.setDate(date.getDate() + value);
-                break;
-            case 'month':
-                date.setMonth(date.getMonth() + value);
-                break;
-            case 'year':
-                date.setMonth(date.getMonth() + value * 12);
-                break;
-        }
-
-        return date.toISOString();
-    }
-
-    let expirationSelect = null;
-    let expirationCustom: string | undefined = undefined;
-
-    $: {
-        $key.expire = expirationSelect === 'custom' ? expirationCustom : expirationSelect;
-    }
 </script>
 
 <WizardStep>
@@ -38,38 +15,6 @@
             placeholder="API Key Name"
             required
             bind:value={$key.name} />
-        <InputSelect
-            bind:value={expirationSelect}
-            options={[
-                {
-                    label: 'Never',
-                    value: null
-                },
-                {
-                    label: '7 Days',
-                    value: incrementToday(7, 'day')
-                },
-                {
-                    label: '30 days',
-                    value: incrementToday(30, 'day')
-                },
-                {
-                    label: '90 days',
-                    value: incrementToday(90, 'day')
-                },
-                {
-                    label: '1 Year',
-                    value: incrementToday(1, 'year')
-                },
-                {
-                    label: 'Custom expiration Date',
-                    value: 'custom'
-                }
-            ]}
-            id="preset"
-            label="Expiration Date" />
-        {#if expirationSelect === 'custom'}
-            <InputDateTime id="expire" label="" bind:value={expirationCustom} showLabel={false} />
-        {/if}
+        <ExpirationInput bind:value={$key.expire} />
     </FormList>
 </WizardStep>
